+++ lilac/src/main/java/com/hartwig/hmftools/lilac/LilacConstants.java
@@ -10,8 +10,8 @@ public class LilacConstants
 {
     public static final String APP_NAME = "Lilac";
 
-    public static final int DEFAULT_MIN_BASE_QUAL = 30;
-    public static int LOW_BASE_QUAL_THRESHOLD = DEFAULT_MIN_BASE_QUAL; // may be adjusted by config or dynamically from median quals
+    public static final byte DEFAULT_MIN_BASE_QUAL = 30;
+    public static byte LOW_BASE_QUAL_THRESHOLD = DEFAULT_MIN_BASE_QUAL; // may be adjusted by config or dynamically from median quals
 
     public static final double LOW_BASE_TRIM_PERC = 0.35;
     public static final double MAX_LOW_BASE_PERC = 0.5;
